+++ examples/basic/src/composed-function/use-tool-or-generate-text-with-custom-prompt-example.ts
@@ -3,7 +3,7 @@ import {
   OpenAIChatMessage,
   OpenAIChatModel,
   Tool,
-  callToolOrGenerateText,
+  useToolOrGenerateText,
 } from "ai-utils.js";
 import dotenv from "dotenv";
 import { z } from "zod";
@@ -35,12 +35,27 @@ dotenv.config();
     execute: async ({ a, b }) => a + b,
   });
 
-  const { tool, result } = await callToolOrGenerateText(
+  const { tool, result } = await useToolOrGenerateText(
     new OpenAIChatModel({ model: "gpt-3.5-turbo" }),
     [multiplyTool, addTool],
-    OpenAIChatFunctionPrompt.forToolsCurried([
-      OpenAIChatMessage.user("What's fourteen to the power of two?"),
-    ])
+    // Instead of using a curried function,
+    // you can also work with the tools directly:
+    (tools) =>
+      OpenAIChatFunctionPrompt.forTools({
+        tools,
+        messages: [
+          OpenAIChatMessage.system(
+            // Here the available tools are used to create
+            // a more precise prompt that reduces errors:
+            `You have ${tools.length} tools available (${tools
+              .map((tool) => tool.name)
+              .join(", ")}).`
+          ),
+          OpenAIChatMessage.user("What's fourteen to the power of two?"),
+          // OpenAIChatMessage.user("What's twelwe plus 1234?"),
+          // OpenAIChatMessage.user("Tell me about Berlin"),
+        ],
+      })
   );
 
   console.log(tool != null ? `TOOL: ${tool}` : "TEXT");