+++ src/model-provider/openai/chat/OpenAIChatPrompt.ts
@@ -1,7 +1,7 @@
 import SecureJSON from "secure-json-parse";
 import z from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
-import { Tool } from "../../../composed-function/call-tool/Tool.js";
+import { Tool } from "../../../composed-function/use-tool/Tool.js";
 import { GenerateJsonPrompt } from "../../../model-function/generate-json/GenerateJsonModel.js";
 import { GenerateJsonOrTextPrompt } from "../../../model-function/generate-json/GenerateJsonOrTextModel.js";
 import { SchemaDefinition } from "../../../model-function/generate-json/SchemaDefinition.js";